0.6.0
- Redid logo
- Decrustified some chunkdata stuff
- Added view distance resizing
- Added a lazy meshing option
- Optimized chunk requests
- Added natural lighting
- Made the random number algorithm slightly more random
- Changed lighting from 12-bit RGB (R4G4B4) to 15-bit RGB (R5G5B5)
- Changed bedrock texture
- Increased world height to 512
- Added cave culling
- Improve config loading

+++ src/game/blocks.c
@@ -61,9 +61,9 @@ void initBlocks() {
                 }
             }
             blockinf[i].data[j].anidiv = atoi(getInfoVarStatic((char*)varcfg->data, "animationdiv", "1", 16));
-            blockinf[i].data[j].light_r = atoi(getInfoVarStatic((char*)varcfg->data, "light_r", "0", 16));
-            blockinf[i].data[j].light_g = atoi(getInfoVarStatic((char*)varcfg->data, "light_g", "0", 16));
-            blockinf[i].data[j].light_b = atoi(getInfoVarStatic((char*)varcfg->data, "light_b", "0", 16));
+            blockinf[i].data[j].light_r = atof(getInfoVarStatic((char*)varcfg->data, "light_r", "0", 16));
+            blockinf[i].data[j].light_g = atof(getInfoVarStatic((char*)varcfg->data, "light_g", "0", 16));
+            blockinf[i].data[j].light_b = atof(getInfoVarStatic((char*)varcfg->data, "light_b", "0", 16));
             blockinf[i].data[j].backfaces = getBool(getInfoVarStatic((char*)varcfg->data, "backfaces", "false", 16));
             freeResource(varcfg);
             #if DBGLVL(1)
